requirements.txt
absl-py==0.8.1
asn1crypto==0.24.0
astor==0.8.0
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
bcrypt==3.1.4
bleach==3.0.2
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.3.9
command-not-found==0.3
cryptography==2.3.1
cupshelpers==1.0
cursor==1.2.0
cycler==0.10.0
decorator==4.4.1
defer==1.0.6
defusedxml==0.5.0
distro-info===0.18ubuntu0.18.04.1
entrypoints==0.2.3
enum34==1.1.6
future==0.16.0
gast==0.3.2
gitdb2==2.0.4
GitPython==2.1.11
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
halo==0.0.18
hdf5storage==0.1.15
httplib2==0.9.2
idna==2.7
imageio==2.6.1
ipykernel==5.1.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.1
Jinja2==2.10
joblib==0.14.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.0
jupyterlab-server==0.2.0
Keras==2.2.5
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
log-symbols==0.0.11
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.0
matplotlib==3.1.2
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
netifaces==0.10.4
networkx==2.4
notebook==5.7.0
numpy==1.18.0
oauth==1.0.1
olefile==0.45.1
opencv-python==3.4.3.18
pandas==0.25.3
pandocfilters==1.4.2
paramiko==2.4.2
parso==0.3.1
pexpect==4.6.0
pgmagick==0.7.5
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.4.1
prompt-toolkit==2.0.5
protobuf==3.10.0
ptyprocess==0.6.0
pyaml==17.12.1
pyasn1==0.4.4
pycairo==1.16.2
pycparser==2.19
pycrypto==2.6.1
pycryptodomex==3.6.6
pycups==1.9.73
Pygments==2.2.0
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
reportlab==3.4.0
requests==2.19.1
requests-unixsocket==0.1.5
requirements-parser==0.2.0
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.4.1
scour==0.36
SecretStorage==2.3.1
Send2Trash==1.5.0
sentry-sdk==0.4.1
simplegeneric==0.8.1
SimpleITK==1.2.4
simplejson==3.13.2
six==1.13.0
smmap2==2.0.4
spell==0.20.3
spinners==0.0.23
ssh-import-id==5.7
system-service==0.3
systemd-python==234
tabulate==0.8.2
tensorboard==1.14.0
tensorboardX==1.9
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
torch==1.3.1
torchvision==0.4.2
tornado==5.1.1
tqdm==4.40.2
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.23
usb-creator==0.3.3
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.4.2
wrapt==1.11.2
xkit==0.0.0
zope.interface==4.3.2
